refactored Login and Signup
