+++ src/components/Copyright.jsx
@@ -1,0 +1,15 @@
+import Link from '@mui/material/Link';
+import Typography from '@mui/material/Typography';
+
+export default function Copyright(props) {
+	return (
+		<Typography variant='body2' color='text.secondary' align='center' {...props}>
+			{'Copyright © '}
+			<Link color='inherit' href='https://www.aidenmayoros.com/'>
+				aidenmayoros.com
+			</Link>{' '}
+			{new Date().getFullYear()}
+			{'.'}
+		</Typography>
+	);
+}
